Changed int to long in id(controller and commands)

+++ VictoryCenter/VictoryCenter.BLL/Commands/TeamMember/DeleteTeamMemberCommand.cs
@@ -3,4 +3,4 @@ using MediatR;
 
 namespace VictoryCenter.BLL.Commands.TeamMember.DeleteTeamMember;
 
-public record DeleteTeamMemberCommand(int Id) : IRequest<Result<Unit>>;
+public record DeleteTeamMemberCommand(long Id) : IRequest<Result<Unit>>;
